gcp_requirements.txt
streamlit>=1.28.0
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0 